Keep built-in editor key bindings in share/binding/builtin
This file is converted to C source code at compile time and embedded
into the program binary. It is also installed as a plain config file for
user reference.

Previously the actual bindings were kept in main.c and then replicated
manually as comments in share/binding/default for user reference. This
new process ensures the reference always matches the bindings by virtue
of sharing the same source code.