Use hybrid variable trick in makefile to reduce build times
This trick combines the advantages of simple and recursive variables, so
as to be expanded only when needed (like a recursive variable) and only
once on first use (like a simple variable).

This avoids expanding the "cc-option" macro (which shells out to the
compiler) repeatedly for each object file.

It's an obscure trick but it reduces the time taken for a full build by
about 20%.