Parse /etc/os-release by just sourcing it in the shell
This was the technique used originally, until it was changed in commit
24b51778f651c198a211544a3d4b0fc7f649654f.

The justifications given in the commit message for 24b51778 are
incorrect. Being able to source /etc/os-release in a shell is explicitly
required by the spec, so it's not "fragile" and it doesn't pollute the
environment because it's sourced in a sub-shell.