Rename "src/ctype.h" to "src/ascii.h"
There is now a standard system header named "ctype.h" in the ISO C11
spec. Having a header with the same name as a system header is a bad
idea, even if it doesn't cause any immediate problems.

It's also really bad form to undef/replace macros defined by the C
standard (e.g. "isupper") -- so all such cases of this have been fixed
by prefixing with "ascii_". This is also more self-documenting when used
throughout the codebase, as it's immediately clear that "ascii_isupper"
is not locale-dependant.