Revert binary search in "allowed_command" function back to linear search
The plan was to gradually convert all similar functions to also use
binary search lookups, but due to the way they're structured and chained
together, it would be a huge and error prone task with minimal benefit.

There's no point in using binary search for only one small lookup table,
so we may just as well revert to using linear search for the sake of
simplicity.

The binary search implementation was also not functioning correctly
anyway.

This reverts commit e46c689441892fda0a535d1f12f532fbd9ea6180.