Remove "bind ^H 'erase-word -s'" from default key bindings
Ctrl+h is sometimes the same as Ctrl+Backspace and sometimes the same as
just Backspace. Binding this to "erase-word -s" by default means that on
some terminals pressing Backspace will delete whole words.

This is obviously a key binding better suited to a personal rc file.
Perhaps in the future it could be set conditionally via a scripting
langage.