Fix compiling on OpenBSD
It seems the GCC shipped with OpenBSD 6.0 doesn't support C11 yet. The
only C11 feature used in this codebase so far is "_Static_assert", which
can just be disabled if not supported.

This commit also removes the use of the "-std=gnu11" flag, since it
causes a build error on OpenBSD and is not actually required. Most of
the compilers that support this option use it by default anyway.