Add "make check-commands" target
This starts the editor and automatically runs a sequence of commands
that produce a specific buffer of text. It then saves the file, quits
the editor and checks that the output is as expected.

This will be the first of many such tests. Unfortunately this style of
testing probably can't be run on CI servers, since they generally don't
have full, interactive terminal emulators.