Adjust warning flags slightly to workaround idiotic Clang behaviour
The "-Wformat=2" option includes the "-Wformat-nonliteral" option, which
works sensibly in GCC.

With Clang, however, it produces a warning for functions that take a
"va_list", which becomes an error when running CI tests with the
"-Werror" option enabled. This is clearly idiotic, as a function that
take a "va_list" cannot be added by accident and necessarily must use a
non-literal format string.

This commit removes "-Wformat=2" from the list of warning flags and
replaces it with "-Wformat -Wformat-security", which is essentially the
same as "-Wformat=2 -Wno-format-nonliteral".