Read terminal input from /dev/tty if stdin is redirected
Previously stdin was assumed to be the controlling tty, but this
assumption doesn't always hold true. For example, running the editor
with the command:

    echo test | dte

...would mean that stdin would be a pipe connected to the stdout of the
echo process.

Before this commit, running such a command would start the editor in a
completely unusable state, as no configuration files would be loaded
and terminal input wouldn't work.