Refuse to open files larger than 256 MiB
The feature set of the editor is tailored to editing source code. No
sane source code file is anywhere close to 256MiB and it doesn't make
sense to open large binary files in dte.

Therefore, to prevent accidentally opening huge files and potentially
making the system unresponsive, we just refuse to open files larger than
256MiB. This can possibly be turned into an configurable option later
if the hard coded limit proves to be a problem.