Specify "terminfo library" as a dependency, instead of "ncurses"
Many platforms use their own implementation of the X/Open Curses spec,
so stating "ncurses" as a dependency is not correct.

The editor also only uses the terminfo database/library and not the
curses API itself. Some implementations split the curses library and the
terminfo library into separate shared object files, so this fact should
be made more clear.