Don't show default, empty buffer if "dte -t tag" opens a file
This is another implementation of commit 8d930a1e8088, which was almost
immediately reverted by commit d77cb1892473 due to a use-after-free bug.

The bug was triggered by using the "-t" flag and also passing a command
to the "-c" flag that would cause the pointer to the empty buffer to be
invalidated (see: is_useless_empty_view() in src/window.c).