Fix DefinitionList handler in docs/pdman.lua
The layout of the "items" parameter recently changed
(see https://github.com/jgm/pandoc/issues/4202).

Apparently this change is just returning to the layout that was used
before it was unintentionally changed by a regression bug.

This commit has been tested with pandoc 2.0.6, which should now be
considered the minimum supported version.