Avoid zero-sized xmemdup in replace_on_line() function in src/search.c
Zero-sized allocations are implementation-defined and thus should be
avoided.

Note: the only way the LineRef could have a size of zero is if the regex
pattern used matched an empty line (e.g. "^"), so just skipping the
xmemdup is safe as there will be no further iterations of the loop.