Revert commit that added `dte -T` command-line flag
This flag was added to allow printing a list of tags matching a prefix
to stdout, for use in shell completion scripts. However, the `readtags`
program that comes with ctags can be used for the same purpose.

The functionality provided by the command:

    dte -T "$PREFIX"

...can more or less be replicated as:

    readtags -p "$PREFIX" | cut -f1 | sort | uniq

The "-T" flag was never documented and thus was not considered part of
the stable command-line "API".

This reverts commit ee7966c85732f8e9a7ac05cbf942a720dabcb4b5.