Only use gperf "-D" flag for extensions.gperf
The "-D" flag prevents blank entries in the table at the cost of some
extra processing.

The other keyword sets have an acceptable number of blank entries, but
the table generated from extensions.gperf is too bloated when the "-D"
flag isn't used -- the strings are too short and too similar.