Remove Perl syntax file
The fact that it doesn't recognize regexp literals makes it basically
useless.

The only way Perl syntax highlighting is ever coming back is if someone
else contributes and maintains it or if some existing definitions for
another editor can be programmatically converted. The syntax of Perl is
just too disgusting to learn purely to maintain a syntax highlighter.