Remove "add built-in support for vt220 terminals" from TODO list
The "vt220" terminfo(5) entry contains many padding specifiers in string
capabilities that must be expanded by tputs(3) (or similar). It's really
not worth reimplenting most of tputs(3) just to add built-in support for
an obsolete terminal.

VT220 *emulators* that don't require padding should really start using a
proper terminfo identifier/entry.

Note: this is just concerning *built-in* support -- the VT220 works fine
when the editor is compiled with terminfo support enabled.