Use new GCC string truncation/overflow warning options
Also use the GCC 8 "nonstring" attribute in a few places to
suppress potential false positives and enable some extra checks.

This commit also fixes a warning about using fputs(3) to print a
"nonstring" char array. Is this case, the member in question was
actually guaranteed to be null-terminated, but using xwrite()
instead of fputs(3) makes more sense regardless.