Clean up string lookup function generator (mk/lperf.lua)
Instead of writing into a string buffer of arbitrary size, it now just
sets stdout(3) to block buffered mode and writes the output directly.

Note: if the script throws an error for some reason and stdout is
redirected to a file, it may produce an invalid, partial file. This
isn't really an issue in practice, since the script is usually called
via the makefile (which has `.DELETE_ON_ERROR` enabled).