Fix off-by-one bug in buf_set_bytes() function
If terminal.repeat_char() was used, it was writing 1 byte too many.
In cases where this was used to pad the ends of lines with spaces,
it was sometimes causing visual glitches due to writing the first
column of the next line.

This commit is perhaps not the cleanest way to fix the problem.
The actual off-by-one is probably in the calculation of the "count"
variable. However, no graphical glitch is visible unless using
terminal.repeat_char(), so this fixes the issue for now.

A minimum threshhold for using terminal.repeat_char() has also been
added. The "REP" control sequence it produces is at least 5 bytes,
so it makes no sense to use it for 5 or fewer repetitions.