Remove "insert-special" command
Arbitrary bytes and codepoints can be inserted into command arguments by
using double-quoted escapes (e.g. "\x1b" or "\u20ac"). This can be used
in conjuction with the "insert" command for much the same effect as
"insert-special". This method can also be used in conjunction with
aliases and key bindings and is much easier to test.

Note: this commit also removes the Ctrl+V key binding from command mode,
which functioned similarly to "insert-special", but inserted into the
command-line instead of the buffer.