Add "strikethrough" terminal attribute
Although this attribute is specified by ECMA-48, many terminals
never implemented it and terminfo has no capability name for it.
Therefore, it has only been implemented for xterm-like terminals
that have built-in support in the editor. Even some of these
terminals will still ignore it or require extra configuration
before it will work.