Set default color mode for xterm back to 8/256
It was changed to true/24-bit color mode when support for that was
added (in commit 8dcbbf6381a4), but many terminals that identify as
"xterm" don't actually implement 24-bit color, including some older
versions of xterm itself.

For now, this commit makes it so that 24-bit colors are only used
when $TERM is set to "xterm-direct". In all other cases, a fallback
from the 8/256 color palette will be used instead.

In the future, 24-bit colors can be enabled by default for other
values of $TERM that work reliably and unambigously. An editor
command to force-enable 24-bit color rendering may also be added.