Prevent infinite recursion in cmd_noeat() function
This could be triggered by using `noeat this` as the default action
in a syntax file, which would execute the exact same action as before
and never make any progress due to the `noeat` command not consuming
anything.

It will always be possible to cause infinite recursion in a badly
authored state machine, but this particular mistake is both easy to
make and easy to prevent.