Add put_control_code() function pointer field to TerminalInfo struct
To allow different implementations, depending on whether or not
terminfo is in use.

The curs_terminfo(3X) documentation states:

> All terminfo strings (including the output of tparm) should be
> printed with tputs or putp.

The implemenetation used in conjunction with terminfo strings now
fully adheres to this requirement.

Before this commit, only the strings listed in the terminfo(5) man
page that indicate padding requirements were printed with tputs().
This list is inaccurate/incomplete, since the vt100 entry adds
padding for the sgr0 string, even though sgr0 is not listed as
requiring padding.