Remove Lua scripting support
Retrofitting this into the existing editor requires too much rewriting
and refactoring, for perhaps little or no gain to most people. I really
don't believe there's a good way to do it properly without redesigning
the editor almost from scratch to make Lua support first class.

There's also the problem of API stability, which is usually the domain
of library projects and is a huge, unpleasant burden for an open source
project.

I may reconsider this decision at some point in the future, but for now
I want to make a release without bloating the tarball with half-baked,
features. In all likelihood though, it probably won't be coming back.