Add basic editorconfig support
For now, the implementation runs the editorconfig(1) command-line
program as a sub-process (if available) and reads the output via a pipe.

To enable globally, the following can be added to ~/.dte/rc:

    set editorconfig true

The implementation may be switched to use the "editorconfig-core-c"
library API eventually (for efficiency), but not until the upstream
project fix their buggy, insecure error handling. Using a buggy library
can crash the entire editor process, whereas the worst a sub-process can
do is crash itself and terminate.

For more information on editorconfig, see:

    https://editorconfig.org/