Fix typo in encoding_char_size() function
The return value should be the number of bytes per codepoint in the
specified encoding, which should clearly be 4 for UTF-32.

This typo has gone unnoticed since the editor was forked from dex,
mostly likely because the value is only used to determine how many
replacement characters to insert for invalid codepoints. In this
particular case, invalid codepoints in UTF-32 files would become 2
inverted question marks instead of 1.