Fix packing of union subtypes in Condition struct
The "cond_char" BitSet within the union is 32 bytes, which should be the
upper limit of the union size.

This commit fixes the size of the `cond_bufis` and `cond_str' fields to
be exactly 32 bytes in total. It also slightly increases the number of
bytes available by using uint8_t instead of size_t to store lengths.

Besides the minor improvements noted above, the main reason for these
changes is to replace incorrect and confusing expressions with simple
and obvious constants.