Optimize fast path of wbuf_write() function slightly
If the number of bytes to be written is larger than the buffer,
it's always going to be larger than the available space -- so the
two conditions can be nested. This means one less check to perform
when the buffer has enough space and one less XFLUSH() in the
generated code.