Just use "-lcurses" to link terminfo library instead of using pkg-config
Using "-lcurses" should work on every sane platform and the X/Open
Curses spec even mentions it specifically.

The commit that introduced pkg-config querying was done to work around
a stupid bug in the Arch Linux ncurses package. The bug was quickly
fixed and Arch is a rolling release, so there's really no reason to
continue using pkg-config unless other platforms require it.

Most linkers are smart enough to prevent over-linking in the case
where terminfo and curses live in separate shared objects.