Rename u_is_space() to u_is_breakable_whitespace() to reflect usage
It's only used to determine break points for commands like word-fwd
and format-paragraph, so calling it "u_is_space" was very misleading.

It now no longer calls u_is_special_whitespace(), since that function
is intended for a completely different purpose, albeit with some
overlap in the range of relevant characters.