Make u_is_unprintable() return true for unassigned Unicode codepoints
This bloats the unprintable[] character table quite a bit, but only
increases the average number of comparisons done by the binary search
by about 3.

In exchange, the editor now reliably renders *all* unprintable
codepoints as <??>, instead of leaving them for the terminal to
display as a fallback character (with unpredictable width and
appearance).