Make `bind` command much more strict when parsing key strings
The caret character can now only be used at the beginning of a key
string and can't be combined with other modifiers.

For example, this is valid:

    bind ^W close

but these are no longer valid:

    bind ^M-W close
    bind M-^W close

The Unix tradition of using the caret to denote "control characters"
never extended to this kind usage and it also just looks ugly and
confusing. The old code from dex that allowed this did so out of
laziness rather than being liberal in what it accepted.