Replace old "inih" INI parser with from-scratch implementation
This eliminates the arbitrary length limits and greedy use of stack
space and is also much faster.

None of the old inih code remains and so the previous copyright and
license header no longer applies. This new code is Copyright Craig
Barnes 2019 and licensed under the GNU General Public License version
2.0.