Change return type of xvsnprintf() and xsnprintf() from int to size_t
The vsnprintf(3) standard library function returns int, but only for
historical reasons. It would make more sense for it to return ssize_t.
With xvsnprintf() the return is never negative, so we can use size_t
instead.

Note: this change isn't for the purpose of increasing the maximum
length. It would be insane to use vsnprintf() with strings anywhere
near INT_MAX bytes. It's more so to allow the compiler to generate
better code and to make the purpose of the return type more obvious.