Move src/util/regexp.c to src/regexp.c
All code in util/ is supposed to be general purpose utility code that
could reasonably be used in other projects with minimal changes. The
main purpose of the wrapper code in regexp.c is to call error_msg() in
case of compilation errors, which means it's inherently coupled to the
editor. The "general purpose" equivalent is just POSIX <regex.h>.