Mention in README.md that GitHub-generated tarballs are NOT immutable
Package maintainers should use the tarballs from the releases page,
which are generated exactly once (at release time) and then archived.
The only time their checksums may ever change are in the case of
actual file corruption, in which case they will be replaced from
backups with bit-identical copies of the originals.