Add "-k" flag to "copy" command, to keep selection after copying
This is to allow running "copy -k" and then some other command on the
same selection.

Most other commands that use the selection necessarily call unselect()
afterwards, because they may modify the size and contents of the text.
The "copy" command is an exception to this though, since it leaves the
text untouched.