Pass EILSEQ instead of errno to fatal_error() if vsnprintf(3) fails
The C standard doesn't seem to require setting errno, so using it
here may produce a completely unrelated error message. This error
should never actually happen in practice, but if for some reason
it does, EILSEQ seems like the most fitting value for an "encoding
error", which is the only error condition mentioned in the standard.