Perform a full-screen clear of the terminal when calling ui_end()
Original issue report copied below:

> When running an external command like e.g. `run sh`, the control
> sequences sent to clear the screen depend on the capabilities of
> the terminal. If it has an alternate screen buffer, the buffers
> are swapped and the entire screen goes back to how it looked
> before dte was started. For terminals that don't have an
> alternate buffer, the image drawn by dte stays in place and
> scrolls out of view only as shell commands are run and produce
> new output.
>
> If the external command being run doesn't immediately print some
> output, the shell prompt gets printed on a line containing part
> of the old screen image drawn by dte and if the shell prompt
> itself doesn't clear it, it'll stay there until something does.
> This is the case with the FreeBSD 12 console and default prompt.
>
> It seems like this is probably a rare occurrence and it has an
> easy workaround (just press Enter once), but it looks ugly and
> confusing when it does happen and should be fixed somehow.

This commit fixes the issue by clearing the entire screen. If
the editor has an alternate screen buffer, then that is the
buffer that gets cleared and the contents of the shell session
before dte was started is retained. Otherwise, if there is only
one screen buffer, it will obviously be the one cleared and the
terminal will just be blank when going into shell mode, or after
quitting the editor. This shouldn't be a problem, since the
contents would otherwise just be the last screen image drawn by
the editor anyway.