Allow substrings in regexp argument of "errorfmt" command to be ignored
By using "_" as the substring name.

This allows parentheses to be used in the regexp pattern without needing
to associate it with a file|line|column|message index. This can be
useful for matching alternate words, e.g. "(word1|word2)", or optional
words, e.g. "(word)?", that don't need to be captured.

Perl regexps have non-capturing groups, e.g. "(?:regexp)", for this
purpose, but dte uses POSIX <regex.h>.