Replace list_for_each_entry() macro with block_for_each()
The only place linked lists are used is in block.h, so there's no need
for a general-purpose macro when something block-specific will do.

This also replaces the container_of() macro with just a simple pointer
cast, since the "node" field is the first member of the Block struct
and thus guaranteed to point to the same location.