Cache simple key bindings and call directly when invoked
Instead of of looking up the command and parsing the arguments each and
every time the key is pressed, this allows the underlying C function to
be called directly.

To be eligible for caching, the following criteria must be met:

* The binding is exactly one command with no syntax errors
* The command is a built-in (not an alias)
* The arguments are valid according to parse_args()
* The arguments do NOT contain variables

This covers all of the default bindings and should also cover the
majority of user bindings.