Mention "ICONV_DISABLE=1" build option in README.md
Note: the motivating use case for this option is reducing the binary
size of statically-linked builds, for use in an initrd image (i.e. a
"recovery" editor).

In most other circumstances, the loss in functionality isn't worth the
difference in file size. For dynamically linked builds, it makes almost
no difference in file size.