Reimplement prev_window() and next_window() to avoid allocations
The previous implementation collected windows into a PointerArray and
then did a linear search for the current window pointer. This was a
totally unnecessary use of dynamic allocation and also caused every
window to be visited twice.

The new implementation does no allocation and only iterates the
windows once -- although there's still room for improvement.