Reset command auto-completion state if command mode is exited
Previously it was only reset when a command was actually run (i.e.
Enter was pressed), but it should also be reset when command mode
is exited without running anything too (i.e. when Esc, ^C or ^G is
pressed). Otherwise pressing Tab with an empty command-line reuses
the previous completion state, which doesn't make much sense.