Make encoding_supported_by_iconv("") return false
The GNU iconv_open(3) implementation accepts an empty string as an
encoding name, which it interprets to mean "the encoding of the current
locale". However, using this in dte (e.g. as "open -e '' file") causes
the status bar to display nothing for the "%E" format specifier. The
minor convenience of this "feature" may be worthwhile in other use
cases, but it makes little sense to allow it here.