Add "mixin" command for dte-syntax(5) files
This is like the existing "include" command, except that it only
tries to load the same file once. The assumption is that such
"mixin" files will define sub-syntaxes that are intended to be
used from several other primary syntax files.

Since all loaded sub-syntaxes are cached, subsequent "mixin" commands
with the same argument will be no-ops and the sub-syntax will just be
served from the cache. The read-once logic is employed to avoid
redundant work and "syntax already exists" warnings.

Note: the "include" command has been retained for use cases that
involve loading the same block of text multiple times, as opposed
to loading (and caching) a shared sub-syntax.