Use "%jd" format specifier instead of "%d" for printing pid_t types
POSIX requires that pid_t is a signed integer type, but it may be
wider than a plain int (e.g. a long int). Using the "%jd" specifier
and casting the pid to intmax_t guarantees that the conversion will
be well defined.

This also changes the buffer allowance for the printed type from 32
bytes to 4 * sizeof(pid_t). Given that POSIX requires CHAR_BIT to be
8, having 4 bytes of buffer per 8 bits of integer width should be
more than enough, whatever the underlying type of pid_t is.