Add "fsync" global option
This reduces unexpected consequences if the system loses power during a
save operation, but potentially increases the length of time the "save"
command has to wait on blocking calls.

It's not entirely clear that using this is worth the extra latency when
editing e.g. files that already have reliable backups in the form of
version control etc. Especially when those files are large or stored on
slow media.

In an ideal world, we'd perhaps use a background thread to complete the
fsync() and rename(), but that entails other difficulties in an
interactive editor.