Re-introduce support for rxvt Ctrl/Shift/Ctrl+Shift key codes
Support for these key codes was removed a while back, as part of a
rewrite of the parser. It was justified at the time, based on the
fact that the format of the escape sequences used is a violation of
the ECMA-48 specification and the rxvt developers are clearly morons
for not just copying the xterm format.

However, a surprising number of people still seem to use rxvt-unicode,
so a basic level of support for it is still a worthwhile thing to have.