Wrap some branch conditions in xterm_parse_key() with unlikely()
This allows the generated code to be ordered in favor of parsing of
valid key sequences and moves the error checking out of the hot path.

If any of the "unlikely" branches are taken, the function just returns
to the caller anyway, so misprediction has a relatively low overhead
in those cases (it can happen at most once per call).