Add "-l" flag to "filter" and "pipe-to" commands
To allow operating on just the current line instead of the entire
buffer when there's no text selected.

An example use case for this is to allow copying to the system
clipboard as well as the internal clipboard in a single keybinding:

    bind ^C 'pipe-to -l xsel -b; copy'
    bind ^X 'pipe-to -l xsel -b; cut'

Without the "-l" flag, there would be a mismatch in the semantics of
the "pipe-to" and "copy" commands when there's no text selection. By
default, "copy" operates on just the current line, whereas pipe-to
operates on the entire buffer.