Replace all uses of LineEndingType enum with bool
The "newline" option retains the same name (for compatibility), as do
the 2 option values, but they're now boolean at the source code level.
Fortunately, the values are exactly the same as before, just with a
slightly different interpretation.

This was done because using an enum for a type with exactly 2 states
is idiotic. That's exactly what bool is for. There are no relevant
line-ending types besides LF and CRLF.