Store GlobalOptions members as unsigned int instead of enum types
Since uint_get() and uint_set() load and store in a way that requires
exactly that type, for alignment purposes.

In most cases, the enum types will resolve to int or unsigned int,
either of which is compatible for load/store purposes, but that
assumption would break if someone compiled with "-fshort-enums"
or something similar.