Add support for "temporary" buffers
These can be created with "open -t" and are never considered
"modified", regardless of any changes made to them. The temporary
status is indicated by "TMP" in the status line and is cleared
if the buffer is saved.

This commit also converts the "show" command to use temporary
buffers, instead of an external pager.

The "-t" flag has been left undocumented for now and is considered
experimental.