Add "-e" (empty) and "-t" (temporary) flags to "wsplit" command
The default "wsplit" behaviour, if used without arguments, has always
been to replicate the current buffer in the new window. The "-e" flag
can be used to prevent this and simply open a new window with an empty
buffer. The "-t" flag does likewise, but also marks it as a temporary
buffer like "open -t" does.