Add term_xnprintf() and use instead of xsnprintf() + term_add_bytes()
This writes directly into the terminal output buffer, instead of
writing to a temporary buffer and then copying to the output buffer.

It takes a manually-specified upper bound for the length, instead
of accurately calculating it like the (now removed) term_sprintf()
function used to do. This works fine in every case where string
formatting is needed, since the upper bound is usually tiny and
not worth the (considerable) extra work required to calculate it.