Remove built-in "$LINES" and "$COLUMNS" special variables
These were added without proper consideration. Many use cases that
would use such a thing actually want the window dimensions rather
than the terminal dimensions.

Something similar may be added again in the future, but it should
be done with a real use case in mind and named in a way that isn't
ambiguous.

This reverts commit d7eb8d344b2e6221feb6faf83cdbdb347d377ab3.