Run "make USE_SANITIZER=1" on GitLab CI test runners
Most of the test runners have old compiler packages that don't support
ASan/UBSan, but in that case the makefile will just print a warning and
continue as normal.

It should still be tried though -- both for runners that do support it
and so that future OS updates start using it as soon as it's available.