Rename NORETURN macro to "noreturn"
This matches the name defined by the ISO C11 <stdnoreturn.h> header.

It also makes it look more like a keyword than a macro, which is a
good thing, since it expands to _Noreturn when compiling in C11 mode
and behaves much more like a keyword than a GCC-style attribute.