Add "match-bracket" command
This moves the cursor to the "bracket" character paired with the one
under the cursor, if any, or otherwise shows an error message. Any
character in the set "[]{}()" is considered a bracket.

Note: the search for the paired character doesn't take language syntax
into account and will count brackets inside strings and comment just
the same as any other context.