Add "select-cursor-char" global option
This allows configuring whether the character "under" the cursor
is included in selections.

Some people may prefer to treat the left edge of the cursor as
the point of selection. Others may use a vertical bar cursor,
where the notion of a character being "under" the cursor doesn't
really exist. For such cases, this option can be set to false.