Allow "ws-error" to highlight trailing whitespace under the cursor
Normally trailing whitespace under the cursor is excluded, because
it's self-evident that there's whitespace there when the cursor is
hovering over it. It can also be annoying when the highlight flashes
for every space typed at the end of a line.

The new behaviour can be enabled with:

    set ws-error all-trailing

The "all-trailing" option is a superset of the old "trailing" option.
Only one or the other needs to be set. Setting both is equivalent to
just setting "all-trailing".