Add "-c" flag to "hi" command
To allow the command to be ignored if any of the specified colors
aren't fully supported by the terminal.

This is to enable something like the following:

    # Set basic comment color for all color-capable terminals
    hi comment cyan

    # Override the above, only if the terminal supports 256 colors
    hi -c comment 245

If this was done without the "-c" flag, the second command would always
override the first and the specified color would simply "degrade" to the
nearest supported one. With the "-c" flag, 8/16 color terminals simply
ignore the second command and keep using the color specified by the
first.