Avoid using "darkgray" for default comment color
Many terminal color schemes seem to make this color completely illegible
against the default background. It makes no sense for 2 basic palette
colors to be so close that using them together makes text unreadable,
but apparently a lot of people still do this anyway. Possibly because
so many apps and libraries call the color "darkblack" or "brightblack",
which is clearly a genius idea.