Fix potential undefined behavior in do_parse_args() function
If the "char" type representation is signed and a command flag greater
than '\x80' was used, it would cause a left shift of a negative value.

For example, the command:

    open "-\xff" GNUmakefile

...would lead to the following UBSan error:

    src/command/args.c:61:54: error: left shift of negative value -1

This commit fixes the issue by using an "unsigned char", which will
then be promoted to an "int" with a guaranteed positive value.