Remove call to begin_change_chain() from macro_play() function
The begin_change_chain() function doesn't work correctly when nested.
This was never a problem before macros were added, but with macros,
it's possible for both macro_play() and a recorded command to start
a chain.

This commit can be reversed again when begin_change_chain() has been
fixed to handle nesting.