Revert previous commit
Most libc implementations of vasprintf(3) just try to write into an
exponentially growing buffer, instead of using C99 semantics to query
the required size. This seems to be the case for glibc on Linux and
all of the BSDs.

There's really no point in calling the libc function if it's typically
worse than the portable fallback we already provide.

This reverts commit 1286479759b2556092996169085b2bc65a756371.