Clean up xvasprintf() function
It was previously implemented on top of a wrapper that was supposed
to mimic the API of the libc vasprintf(3) function. However, as noted
in the previous commit, the libc function is usually some untouched,
garbage code from the late 90s. So there's really no point in trying
to make it understandable or implementable in terms of vasprintf(3).