Reimplement history_add() using a HashMap of doubly-linked entries
This allows duplicates to be found and re-inserted at the end of the
history in average constant time, as opposed to linear time with the
previous approach.

The previous approach also meant that history_load() was quadratic
in the number of entries due to history_add() being called n times
and walking the array of n entries for each call. This was limited
to a small n by the max_entries limit, but it needed to be fixed
in order to allow the limit to be increased (or made configurable).