Add free_syntaxes() function and call from test/main.c
Syntaxes are usually cached for the entire life of the program and
hence don't need to be explicitly freed. The only time they do get
freed by the main program is when they contain errors. However, we
want to ensure the various free_*() routines for all the internal
data structures get exercised by ASan and UBSan, so we explicitly
free all cached syntaxes just before the test binary exits.