Remove mention of "kitty's extended keyboard protocol" from CHANGELOG.md
Support was added for the "old" keyboard protocol, which now looks like
it will be superceded by another, more capable protocol:

* https://github.com/kovidgoyal/kitty/issues/3248
* https://sw.kovidgoyal.net/kitty/keyboard-protocol.html

To avoid confusion, it's probably best to just not mention it in the
changelog at all, since the documentation for the old protocol seems to
have disappeared. Support for the new protocol will likely be added at
some point.