Use uint_least64_t instead of uint64_t for CommandArgs::flag_set field
The uint64_t type is considered optional, even by the POSIX spec and
even on systems that do have a working, 64-bit, unsigned type. This is
because the {u,}intN_t types have additional requirements, besides just
the bit width, which some systems may not be able to fully satisfy.

This is a mega nitpick, since every operating system we care about does
have a uint64_t type, but in this case there's no downside to using the
(theoretically) more portable of the two.