Move alias handling code back to src/command/alias.c
We eventually want to add Command/CommandArgs caching, in a similar
fashion to that already done for KeyBinding. When that happens, the
code will be quite a bit more complex.

Also, the point of putting command-related code in src/command/ is
that it could, in theory, be used in another project with minimal
changes. If src/command/run.c calls find_alias(), the rest of the
code for adding/removing/finding aliases should also be in the same
directory.

This more or less reverts commit 7d8fdbdf4a81a.