Optimize parse_args() function by making use of __builtin_expect()
This makes very little difference to interactive use, where commands
only need to be parsed as fast as the user can input them.

However, it is very much worthwhile when parsing config files, which
can contain thousands of commands, typically with 100% of parse_args()
calls taking the ARGERR_NONE branch.