Simplify ascii_isdigit() implementation
Since all ASCII digits are contiguous, this function can be implemented
as a simple range check instead of a table lookup.

In the case of buf_parse_uintmax(), this allows the call to
ascii_isdigit() inside the while loop to be compiled as e.g.:

    sub edx, 48
    cmp edx, 9

...instead of:

    test BYTE PTR ascii_table[rdx], 2

This reduces the latency from 6 to 2.