Wrap some more error handling branches with unlikely()
In syntax files, a single error causes loading to fail, so errors
should never happen in stable syntax files and the worst that can
happen in files with errors is one unlikely branch being taken.

So the downsides of doing this are almost non-existent and the
upside is more cache efficient code generation.