Enable "modifyOtherKeys" mode at startup
This mode originates from xterm and is supported by several other
terminals as well as tmux (since version 3.2).

The CSI sequences to enable/disable the mode are always emitted,
regardless of the value of $TERM, since they should be be ignored by
terminals that don't recognize them and also accepted by terminals
that *do* recognize them but are spoofing $TERM for whatever reason.

See also:

 https://invisible-island.net/xterm/manpage/xterm.html#VT100-Widget-Resources:modifyOtherKeys