Add "exec-msg" command
This is undocumented for now and may be changed and/or renamed
before the next release.

Example usage:

    tag main
    exec-msg fzf
    git-grep -w Command
    exec-msg fzf

Note that the second example is lacking some useful information due
to the way the "basic" errorfmt regex is specified.

Normally, when using "msg -n", there's no need to show the matched
string because the cursor jumps to the relevant location. However,
this isn't the case when showing all messages in a list. This could
be improved in the future, either by changing the "basic" errorfmt
to extract the full message or by adding a flag to allow tweaking
the displayed format depending on the context.